[TRL-430] refactor: JwtProvider DIP 적용 및 패키지 이름 변경
