--- src/main/java/com/cosain/trilo/auth/infra/TokenAnalyzer.java
@@ -1,11 +1,0 @@
-package com.cosain.trilo.auth.infra;
-
-import com.cosain.trilo.auth.infra.jwt.UserPayload;
-
-public interface TokenAnalyzer {
-    boolean validateToken(String token);
-    boolean isValidToken(String authorizationHeader);
-    UserPayload getPayload(String authorizationHeader);
-    Long getUserIdFromToken(String token);
-    Long getTokenRemainExpiryFrom(String token);
-}
