--- src/main/java/com/cosain/trilo/auth/infra/TokenProvider.java
@@ -1,7 +1,0 @@
-package com.cosain.trilo.auth.infra;
-
-
-public interface TokenProvider {
-    String createAccessTokenById(final Long id);
-    String createRefreshTokenById(final Long id);
-}
