+++ src/main/java/com/cosain/trilo/trip/presentation/schedule/ScheduleDeleteController.java
@@ -1,10 +1,9 @@
 package com.cosain.trilo.trip.presentation.schedule;
 
-import com.cosain.trilo.auth.infra.jwt.UserPayload;
+import com.cosain.trilo.auth.infra.token.UserPayload;
 import com.cosain.trilo.auth.presentation.Login;
 import com.cosain.trilo.auth.presentation.LoginUser;
 import com.cosain.trilo.trip.application.schedule.service.ScheduleDeleteService;
-import com.cosain.trilo.user.domain.User;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
