+++ src/main/java/com/cosain/trilo/trip/presentation/schedule/ScheduleMoveController.java
@@ -1,6 +1,6 @@
 package com.cosain.trilo.trip.presentation.schedule;
 
-import com.cosain.trilo.auth.infra.jwt.UserPayload;
+import com.cosain.trilo.auth.infra.token.UserPayload;
 import com.cosain.trilo.auth.presentation.Login;
 import com.cosain.trilo.auth.presentation.LoginUser;
 import com.cosain.trilo.trip.application.schedule.service.ScheduleMoveService;
@@ -9,7 +9,6 @@ import com.cosain.trilo.trip.application.schedule.dto.ScheduleMoveResult;
 import com.cosain.trilo.trip.application.schedule.dto.factory.ScheduleMoveCommandFactory;
 import com.cosain.trilo.trip.presentation.schedule.dto.request.ScheduleMoveRequest;
 import com.cosain.trilo.trip.presentation.schedule.dto.response.ScheduleMoveResponse;
-import com.cosain.trilo.user.domain.User;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
