+++ src/main/java/com/cosain/trilo/trip/presentation/trip/TripImageUpdateController.java
@@ -1,12 +1,11 @@
 package com.cosain.trilo.trip.presentation.trip;
 
-import com.cosain.trilo.auth.infra.jwt.UserPayload;
+import com.cosain.trilo.auth.infra.token.UserPayload;
 import com.cosain.trilo.auth.presentation.Login;
 import com.cosain.trilo.auth.presentation.LoginUser;
 import com.cosain.trilo.common.file.ImageFile;
 import com.cosain.trilo.trip.application.trip.service.TripImageUpdateService;
 import com.cosain.trilo.trip.presentation.trip.dto.response.TripImageUpdateResponse;
-import com.cosain.trilo.user.domain.User;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
