+++ src/main/java/com/cosain/trilo/trip/presentation/trip/TripPeriodUpdateController.java
@@ -1,6 +1,6 @@
 package com.cosain.trilo.trip.presentation.trip;
 
-import com.cosain.trilo.auth.infra.jwt.UserPayload;
+import com.cosain.trilo.auth.infra.token.UserPayload;
 import com.cosain.trilo.auth.presentation.Login;
 import com.cosain.trilo.auth.presentation.LoginUser;
 import com.cosain.trilo.trip.application.trip.service.TripPeriodUpdateService;
@@ -8,7 +8,6 @@ import com.cosain.trilo.trip.application.trip.dto.TripPeriodUpdateCommand;
 import com.cosain.trilo.trip.application.trip.dto.factory.TripPeriodUpdateCommandFactory;
 import com.cosain.trilo.trip.presentation.trip.dto.request.TripPeriodUpdateRequest;
 import com.cosain.trilo.trip.presentation.trip.dto.response.TripPeriodUpdateResponse;
-import com.cosain.trilo.user.domain.User;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
