+++ src/test/java/com/cosain/trilo/unit/auth/application/AuthServiceTest.java
@@ -1,6 +1,7 @@
 package com.cosain.trilo.unit.auth.application;
 
 import com.cosain.trilo.auth.application.AuthService;
+import com.cosain.trilo.auth.application.JwtProvider;
 import com.cosain.trilo.auth.application.OAuthProfileRequestService;
 import com.cosain.trilo.auth.application.dto.KakaoLoginParams;
 import com.cosain.trilo.auth.application.dto.LoginResult;
@@ -8,10 +9,7 @@ import com.cosain.trilo.auth.application.dto.OAuthLoginParams;
 import com.cosain.trilo.auth.application.dto.ReIssueAccessTokenResult;
 import com.cosain.trilo.auth.domain.repository.TokenRepository;
 import com.cosain.trilo.auth.infra.OAuthProfileDto;
-import com.cosain.trilo.auth.infra.TokenAnalyzer;
-import com.cosain.trilo.auth.infra.TokenProvider;
 import com.cosain.trilo.auth.presentation.dto.RefreshTokenStatusResponse;
-import com.cosain.trilo.common.exception.NotExistRefreshTokenException;
 import com.cosain.trilo.common.exception.NotValidTokenException;
 import com.cosain.trilo.user.application.UserService;
 import com.cosain.trilo.user.domain.AuthProvider;
@@ -34,9 +32,7 @@ class AuthServiceTest {
     @Mock
     private TokenRepository tokenRepository;
     @Mock
-    private TokenAnalyzer tokenAnalyzer;
-    @Mock
-    private TokenProvider tokenProvider;
+    private JwtProvider jwtProvider;
     @Mock
     private UserService userService;
     @Mock
@@ -47,10 +43,10 @@ class AuthServiceTest {
     @Test
     void ACCESS_토큰_재발급(){
         Long id = 1L;
-        given(tokenAnalyzer.validateToken(any())).willReturn(true);
+        given(jwtProvider.isValidRefreshToken(any())).willReturn(true);
         given(tokenRepository.existsRefreshTokenById(any())).willReturn(true);
-        given(tokenAnalyzer.getUserIdFromToken(any())).willReturn(id);
-        given(tokenProvider.createAccessTokenById(anyLong())).willReturn(ACCESS_TOKEN);
+        given(jwtProvider.getUserIdFromToken(any())).willReturn(id);
+        given(jwtProvider.createAccessToken(anyLong())).willReturn(ACCESS_TOKEN);
 
         ReIssueAccessTokenResult result = authService.reissueAccessToken(any());
         Assertions.assertThat(result.getAccessToken()).isEqualTo(ACCESS_TOKEN);
@@ -58,24 +54,17 @@ class AuthServiceTest {
 
     @Test
     void 접근토큰_재발급시_재발급_토큰이_유효하지_않다면_에러를_발생시킨다(){
-        given(tokenAnalyzer.validateToken(any())).willReturn(false);
+
+        given(jwtProvider.isValidRefreshToken(any())).willReturn(false);
 
         assertThatThrownBy(() -> authService.reissueAccessToken(any())).isInstanceOf(NotValidTokenException.class);
-    }
-
-    @Test
-    void 접근토큰_재발급시_토큰_저장소에_재발급_토큰이_존재하지_않는다면_에러를_발생시킨다(){
-        given(tokenAnalyzer.validateToken(any())).willReturn(true);
-        given(tokenRepository.existsRefreshTokenById(any())).willReturn(false);
-
-        assertThatThrownBy(() -> authService.reissueAccessToken(any())).isInstanceOf(NotExistRefreshTokenException.class);
     }
 
     @Test
     void 토큰_정보_생성(){
 
         // given
-        given(tokenAnalyzer.validateToken(any())).willReturn(true);
+        given(jwtProvider.isValidRefreshToken(any())).willReturn(true);
 
         // when
         RefreshTokenStatusResponse dto = authService.createTokenStatus(any());
@@ -87,7 +76,7 @@ class AuthServiceTest {
     @Test
     void 토큰_정보_생성시_토큰이_유효하지_않는다면_false를_반환한다(){
         // given
-        given(tokenAnalyzer.validateToken(any())).willReturn(false);
+        given(jwtProvider.isValidRefreshToken(any())).willReturn(false);
 
         // when
         RefreshTokenStatusResponse dto = authService.createTokenStatus(any());
@@ -99,7 +88,7 @@ class AuthServiceTest {
     @Test
     void 로그아웃(){
         // given
-        given(tokenAnalyzer.getTokenRemainExpiryFrom(any())).willReturn(100000L);
+        given(jwtProvider.getTokenRemainExpiry(any())).willReturn(100000L);
 
         // when
         authService.logout("Bearer accessToken", anyString());
@@ -124,8 +113,8 @@ class AuthServiceTest {
 
         given(userService.createOrUpdate(any(OAuthProfileDto.class))).willReturn(1L);
         given(OAuthProfileRequestService.request(any(OAuthLoginParams.class))).willReturn(oAuthProfileDto);
-        given(tokenProvider.createAccessTokenById(any())).willReturn(ACCESS_TOKEN);
-        given(tokenProvider.createRefreshTokenById(any())).willReturn(REFRESH_TOKEN);
+        given(jwtProvider.createAccessToken(any())).willReturn(ACCESS_TOKEN);
+        given(jwtProvider.createRefreshToken(any())).willReturn(REFRESH_TOKEN);
 
         // when
         LoginResult loginResult = authService.login(KakaoLoginParams.of(code, redirectUri));
